create child component
